requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
mysqlclient==2.2.7
propcache==0.3.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.5.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
